pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "monarch-phenology"
version = "0.1.0"
description = "Build a monarch phenology dataset from iNaturalist observations and AI image classification."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
dependencies = [
  "httpx>=0.27.2",
  "psycopg[binary,pool]>=3.2.9",
  "pydantic>=2.10.6",
  "python-dotenv>=1.0.1",
  "typer>=0.15.1",
]

[project.scripts]
monarch = "monarch_phenology.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/monarch_phenology"]

